requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
sqlalchemy==2.0.18
pydantic==2.0.2
email-validator==2.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
alembic==1.11.1
pytest==8.3.5
httpx==0.24.1
